move multiline parser to abc-class
